Move api webapp creation to product-is and add sample webapp
