+++ components/org.wso2.carbon.identity.api.dispatcher.core/src/main/java/org/wso2/carbon/identity/api/dispatcher/core/JsonProcessingExceptionMapper.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wso2.carbon.identity.api.dispatcher.core;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.wso2.carbon.identity.api.user.common.error.ErrorDTO;
+import org.wso2.carbon.identity.api.user.common.error.ErrorResponse;
+
+import javax.ws.rs.core.HttpHeaders;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+
+/**
+ * Handles exceptions when an incorrect json requests body is received.
+ * Sends a default error response.
+ */
+public class JsonProcessingExceptionMapper implements ExceptionMapper<JsonProcessingException> {
+
+    private static final Log log = LogFactory.getLog(JsonProcessingExceptionMapper.class);
+    public static final String ERROR_CODE = "UE-10000";
+    public static final String ERROR_MESSAGE = "Invalid Request.";
+    public static final String ERROR_DESCRIPTION = "Provided request body content is not in the expected format.";
+
+    @Override
+    public Response toResponse(JsonProcessingException e) {
+        if (log.isDebugEnabled()) {
+            log.debug("Provided JSON request content is not in the valid format:", e);
+        }
+        ErrorDTO errorDTO = new ErrorResponse.Builder().withCode(ERROR_CODE)
+                .withMessage(ERROR_MESSAGE)
+                .withDescription(ERROR_DESCRIPTION).build();
+        return Response.status(Response.Status.BAD_REQUEST)
+                .entity(errorDTO)
+                .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON).build();
+    }
+}
