+++ sample/org.wso2.carbon.identity.api.dispatcher.sample/src/main/webapp/META-INF/webapp-classloading.xml
