+++ sample/org.wso2.carbon.identity.api.dispatcher.sample/src/main/webapp/WEB-INF/web.xml
